+++ packages/connect/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @databite/connect
 
+## 4.0.0
+
+### Major Changes
+
+- Switched data providers for connection stores. Switched scheduler adapters for databites own internal scheduler to be default always
+
+### Patch Changes
+
+- Updated dependencies
+  - @databite/types@6.0.0
+
 ## 3.0.1
 
 ### Patch Changes