+++ packages/example-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-server
 
+## 2.0.0
+
+### Major Changes
+
+- Switched data providers for connection stores. Switched scheduler adapters for databites own internal scheduler to be default always
+
+### Patch Changes
+
+- Updated dependencies
+  - @databite/connectors@5.0.0
+  - @databite/engine@4.0.0
+  - @databite/server@3.0.0
+  - @databite/types@6.0.0
+
 ## 1.0.1
 
 ### Patch Changes